[VMCL] Add C++ Compute Abstraction example.
With this patch we will add the first Accelerator example. The example
uses a templated class and function with variable arguments to present
a high-level Compute Accelerator abstraction in C++. To support a
high-level abstraction, the details of Accelerator management is handled
by the vmaccelmgr_svr. When available, a vmcl_svr can register itself
with the vmaccelmgr_svr, and a client (e.g. basic_compute) can query
the vmaccelmgr_svr for the location of an available vmcl_svr that
matches the workload requirements.

Example tested on MacOS 10.13:

   Shell 1 - build $ bin/vmaccelmgr_svr
   Shell 2 - build $ bin/vmcl_svr en0 127.0.0.1
   Shell 3 - build $ examples/basic_compute

In the example above, the vmcl_svr registers itself over the "en0"
network interface with vmaccelmgr_svr located at "127.0.0.1".
basic_compute is hard coded to look at "127.0.0.1" for vmaccelmgr_svr
so it auto-resolves and connects to the vmcl_svr.

Since Accelerators have deferred execution, the C++ compute abstraction
makes use of referenced objects via std::shared_ptr. A shared_ptr's
ownership is incremented into a vmaccel::ref_object, then given to the
VMCL client framework. The type checking forces std::shared_ptr to
avoid dangling allocations that are still pending on the Accelerator.

Testing Done: vmcl_svr with basic_compute on the same system.