[VMCL] Free the resulting allocation from vmcl_surfacemap_1.
With this patch we will free the resulting allocation from
vmcl_surfacemap_1. The return value of result_2 contains a temporary
memory allocation that is meant to store the results of the map
operation. Once the caller is done with the operation, it can free
the memory.

This introduces the following memory model for VMAccel Clients,
where inheritance of content is from left-to-right and "->" denotes
an abstraction boundary:

  std::shared_ptr -> ref_object<std::shared_Ptr> -> RPC Call
  ...
  RPC Results -> temp allocation -> caller copies content and frees

The above allows for the memory backing std::shared_ptr not to be
freed until the RPC Call consumes the contents. It also allows
for the content retrieved over the wire to be available to the
caller until the caller can copy or consume the content, then free
the temporary allocation.

For VMAccel Servers the following is true:

  RPC Call -> temporary allocation -> server copies content and frees
  ...
  global memory -> RPC Results -> ... -> deferred re-use of memory

The above Server memory model is a result of embedding pointers in
return structures where control is handed back to the caller without
lookback to cleanup the temporary allocation once the caller has
consumed it.

Testing Done: vmcl_svr + basic_compute