[VMAccel] Add move constructors to the allocator types.
With this patch we will add move constructors to the allocator types.
If an allocation is requested in the RPC protocol, the result struct
may reference a dynamic array that needs to be communicated to the
client. The RPC protocol does not give an opportunity for the server
code to free the associated memory, so a deep copy is not an option.
To handle this scenario, the server will track allocated objects in
the vmaccel_allocator allocated vector.

If there are any VMAccelObject(s) in the allocated vector at
destruction time, the Destructor will be called for the associated
allocator type. Each Destructor handles any associated allocations
and frees the memory.

Testing Done: valgrind test_vmaccel_allocator_desc_test