[VMCL] First pass at Operation Objects.
With this patch we will add support for Operation Objects.
Operation Objects provide an explicit lifetime via the scope of
the Operation Object variable. The initial Operation Object
implementation will handle quiescing of Operation Objects at
their destruction, thus updating all contents for an operation's
bindings. To support this functionality, a clcontext class will
be created via a constructor, and if any resource is unavailable
the constructor will throw an exception.

Testing Done: async_compute on localhost with OpenCL 1.2 SDK.