[VMCODEC] Adjust vmcodec context allocation to take an individual enumerant.
With this patch we will change the vmcodec context allocation to take an
individual enumerant for the intended codec. Each codec has different
state in each of their processes, therefore prevent the ambiguity of
what state is being tracked by limiting the context to one codec config.

Testing Done: build and make pre-checkin