[VMCODEC][VMCL] Move decode/encode accelerator types into codec caps.
With this patch we will move the decode/encode accelerator types into
the codec caps. This allows for a unified codec API that exposes
either decode, encode, or both. To support codec acceleration, a
throughput field was added to the codec context descriptor structure
to determine if an accelerator can support the codec's throughput
requirements. This is more for video codecs which may have low power
clients.

Also included are general cleanup of formatting and renaming of the
test directory to tests.

Testing Done: build and make pre-checkin