[VMAI] Add the basis for an AI acceleration abstraction.
With this patch we will implement the basis for the abstraction
we will use with AI acceleration. The first class is a classifier,
which will act in a stateless fashion to build a model based upon
input. The abstraction is high-level to leverage a Compute
Accelerator and put retention of previous revisions of the model
as a top priority.

Testing Done: build, make pre-checkin