[VMAccel] Move C++ protocol from UDP to TCP.
With this patch we will move the C++ protocol from UDP to TCP.
UDP has a packet limit size ~ 65K, so we need to move to TCP
to avoid failures when sending large amounts of data over the wire.
To test this functionality, we will add a new memory benchmark,
membench, that uses a compute kernel to saturate an accelerator's
compute resources.

membench does a copy from one dataset to another as follows:

   B += A

membench command line usage is as follows:

  membench -m <# rows> -n <# cols> -l <# chunks> -i <# iterations>
           --memoryPoolA <pool enum> --memoryPoolB <pool enum>
           --kernelFunc <kernel function enum>

TODO:
- Add functionality for subdivision of large transfers
- Add functionality for non-blocking asynchronous readbacks

Testing: build, memtest, make pre-checkin