[VMCL] Use upload paths for upload_surface action
With this patch we will use the image upload paths for the upload_surface
action. This requires us to make clEnqueueWriteBuffer blocking, so the
pointer given is not freed before uploading the data. The change reduces
the overall runtime for the UPLOAD_A kernelFunc from 30,722 ms to 15,388 ms.

Testing Done: build, membench_local, membench