[VMCL] Add experimental data stream functionality
With this patch we will add experimental data stream functionality.
This allows for the data to be sent asynchronously to the RPC channel
and therefore unblock RPC communication and allows us to saturate the
interconnect.

This does come with some caveats:
* No more than one client thread can use a stream at a given time.
  This is enforced by the IdentifierDB in vmaccel_streams.c. This is
  to avoid out-of-order information in the stream, which may result
  in corruption of the working set on the server.
* Only one thread can instantiate RPC and Stream requests. This is to
  avoid out-of-order content landing on the server and corrupting
  the working set.
* Currently only one generation is allowed on the server at a given
  time, this is to avoid memory pressure on the server when the GPU
  is the bottleneck. See VMACCEL_MAX_ACTIVE_GENERATIONS.

One can enable the functionality by setting ENABLE_VMCL_STREAM_SERVER
to "1". It will spawn a TCP socket on port 5100 by default. For a stream
i, it will spawn a TCP socket on port 5100+i.

TODO:
* More validation of the stream data and bounds checking
* Timestamp or token validation to allow out-of-order receive on the
  server to reject Dispatch

Experiments were run as follows:

  ./membench --kernelFunc -3
  ./membench_local --kernelFunc -3

Results were as follows for 13,690,728,792 bytes on vmxnet3:
* membench_local (native): 1283 ms -> 10.6 GBytes/sec
* iperf (localhost): 5.5 GBytes/sec
* membench no-streams (localhost): 15406 ms -> 0.888 GBytes/sec
* membench streams 1x1 (localhost): 2761 ms -> 4.5 GBytes/sec

Testing Done: build, make pre-checkin, membench, membench_local