[VMCL] Change to make generation value to track client/server consistency
With this patch we will change the usage of the generation value to provide
client/server data consistency checks between the main RPC thread and the
data stream. This allows for an application to detect when it has created
a race condition in the workload.

To retain the ability to overlap surface updates with compute, we now
use the instance value to explicitly address different copies of a surface
form the client. This allows for experimenting with throughput when an
update is blocked by a compute operation reference.

Testing Done: build, pre-checkin, membench, membench_local