[VMCL] Fix stream resource unavailable retry and add statistics
With this patch we will introduce a retry count and linear backoff
algorithm to allow an out-of-band stream upload to take effect on
the server for use in a dispatch. This linear backoff algorithm is
enforced with pthread_yield and usleep(retryCount * 1000). The
backoff algorithm assumes the stream thread is making progress and
is fairly scheduled. Furthermore, the backpressure is worst case
linear to the size of the working set.

If the backpressure is quadratic or worse, the linear backoff
algorithm will fail. One can debug this scenario via:

* DEBUG_STATISTICS
* DEBUG_COMPUTE_OPERATION
* DEBUG_SURFACE_CONSISTENCY

Testing Done: build, pre-checkin, membench_local, membench