[VMCL] Add counter stat and rename time stat
With this patch we will add a counter stat and rename the time stat.
Included in this change is a new stat to track the number of times
a dispatch resulted in resource unavailable. A resource unavailable
during dispatch represents backpressure from the server showing
that one or more dependencies have not landed on the server. This
is typical for large surfaces transferred with the data stream
functionality.

Testing Done: build, pre-checkin, membench