[VMCL] Add multiple queue support and quiesce test.
With this patch we will add multiple queue per sub-device support
to the VMCL abstraction. This allows for asynchronous upload and
download of surfaces. Multiple queue support is demonstrated by the
new membench epsilon jump kernel.

The epsilon jump kernel will exit out of the inner loop when a
semaphore value equals 1. The multiple queue support is required
to update the semaphore value asynchronously to the execution of
the compute kernel, without SVM functionality. Once updated, a time
is taken to measure the time it takes to update semaphore and
download the working set in FIFO order to the execution of the
compute kernel (a.k.a. the time to quiesce the accelerator).

Sample usage is as follows:

./membench_local -i 1 -m 4 -n 4 -k 16777216 --kernelFunc 9 --memoryPoolS 256 --epsilonDelayMS 1000

Testing Done: pre-checkin, membench_local