[VMCL] Add support for OpenCL 3.0 and Image Fill
With this patch we will add support for OpenCL 3.0 and Image
Fill. OpenCL 3.0 is now an allowed platform and will not result
in no platforms. Along with this support, we add an Image Fill
function to memset a memory range with a 32-bit value. The
size of the region and surface must be 16 byte aligned.

One can benchmark the Image Fill functionality with membench
as follows:

  ./membench_local --kernelFunc -4

Testing Done: Build, membench_local OpenCL 3.0 with CUDA 11.4