[VMCL] Remove remaining VMAccel_IsLocal references
In local mode, the pointers are shared between the client and the
server binaries as they are in the same process. Local execution
in a Hybrid mode is not a compile time decision so we adjust the
checks in the vmwopencl backend to only free RPC memory when requested
by the client.

Testing Done: build, vmcl_membench (local), vmcl_rpc_membench (remote)