[VMCL] Add hybrid compute build target and test
With this patch we will add the hybrid compute target to enable
local and remove compute. As an example, add the hybrid_compute
example source that runs a Matrix2D Add kernel on the local
accelerator and then migrates the working set to the remote
accelerator and dispatches the same Matrix2DAdd kernel.

Testing Done: build, hybrid_compute