[VMAccel] Provide thread safe vmaccel and vmcl APIs
The RPC APIs for vmaccel and vmcl use a static variable for backing return
structures. When in a multithreaded scenario, the static variable may get
clobbered while in use by another thread. To avoid this, add a service mutex
which is a global mutex for any vmaccel enabled process using vmaccel.hpp
classes. The mutex is initialized in the vmaccel_utils_poweron_svc function.

The side effect of this mutex is reduced parallelism due to long running RPCs,
however most data transfer APIs are O(1) except for the map, unmap, upload,
and download functions. If this becomes a problem, one can use the data
stream code to defer working set residency to a poll/retry interaction.

Testing Done: build