[VMAccel] Separate the mutex domains into compute, data, and state
With the last change, we added a mutex to provide thread safety for
the RPC APIs. The mutex would effectively make all APIs O(n) in the
worst case, since a data transfer could block the application from
a state mutator API. To allow for more parallelism without data
streams, we split the service mutex into three domains:

1) compute
2) data
3) state

Data is expected to be O(n) in runtime and storage, but compute and
state are expected to be O(1). This allows the RPC functionality to
have a meaningful timeout value for each of the functionalities, where
a generic metric would be unable to discern between a communication
timeout and execution timeout.

Testing Done: build